Reland "Add format-specifier warnings to SkDebugf."
This is a reland of e58831cd9578805634b986b434a20757277b3a36

Original change's description:
> Add format-specifier warnings to SkDebugf.
>
> This CL fixes up many existing format-specifier violations in Skia.
> Note that GCC has a warning for formatting nothing, so existing calls to
> `SkDebugf("")` have been removed, or replaced with `SkDebugf("%s", "")`.
> These were apparently meant to be used as a place to set a breakpoint.
>
> Some of our clients also use SkDebug with bad format specifiers, so this
> check is currently only enabled when SKIA_IMPLEMENTATION is true.
>
> Bug: skia:12143
> Reviewed-on: https://skia-review.googlesource.com/c/skia/+/420902
> Auto-Submit: John Stiles <johnstiles@google.com>
> Commit-Queue: Brian Osman <brianosman@google.com>

Bug: skia:12143

+++ src/gpu/GrCaps.cpp
@@ -426,8 +426,8 @@ GrSwizzle GrCaps::getReadSwizzle(const GrBackendFormat& format, GrColorType colo
         if (colorType == GrColorType::kRGB_888x || colorType == GrColorType::kRGBA_8888) {
             return GrSwizzle::RGBA();
         }
-        SkDEBUGFAILF("Illegal color type (%d) and compressed format (%d) combination.", colorType,
-                     compression);
+        SkDEBUGFAILF("Illegal color type (%d) and compressed format (%d) combination.",
+                     (int)colorType, (int)compression);
         return {};
     }
